Implementa o domain service para cozinhas

+++ src/main/java/com/github/hugovallada/algafoodapi/infra/repository/CozinhaRepositoryImpl.java
@@ -1,19 +1,19 @@
 package com.github.hugovallada.algafoodapi.infra.repository;
 
-import java.util.List;
-
-import org.springframework.stereotype.Repository;
-
 import com.github.hugovallada.algafoodapi.domain.model.Cozinha;
 import com.github.hugovallada.algafoodapi.domain.repository.CozinhaRepository;
-
 import jakarta.persistence.EntityManager;
 import jakarta.transaction.Transactional;
+import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+import java.util.Objects;
 
 @Repository
-public class CozinhaRepositoryImpl implements CozinhaRepository {
+class CozinhaRepositoryImpl implements CozinhaRepository {
 
-    private EntityManager entityManager;
+    private final EntityManager entityManager;
 
     public CozinhaRepositoryImpl(EntityManager entityManager) {
         this.entityManager = entityManager;
@@ -39,8 +39,11 @@ public class CozinhaRepositoryImpl implements CozinhaRepository {
 
     @Override
     @Transactional
-    public void remover(Cozinha cozinha) {
-        final var cozinhaParaRemover = buscar(cozinha.getId());
+    public void remover(Long id) {
+        final var cozinhaParaRemover = buscar(id);
+        if (Objects.isNull(cozinhaParaRemover)) {
+            throw new EmptyResultDataAccessException(1);
+        }
         entityManager.remove(cozinhaParaRemover);
     }
